+++ src/barbrawl.js
@@ -6,7 +6,7 @@
 import { extendBarRenderer, redrawBar } from "./module/rendering.js";
 import { extendTokenConfig } from "./module/config.js";
 import { extendTokenHud } from "./module/hud.js";
-import { registerSettings } from "./module/settings.js";
+import { getDefaultResources, registerSettings } from "./module/settings.js";
 import { prepareUpdate } from "./module/synchronization.js";
 import { refreshBarVisibility } from "./module/api.js";
 
@@ -88,23 +88,35 @@ Hooks.on("updateToken", function (doc, changes) {
     if (token.hasActiveHUD) canvas.tokens.hud.render();
 });
 
-/** Hook to initialize tokens with default bars. */
+/** Hooks to initialize tokens and actors with default bars. */
 Hooks.on("preCreateToken", function (doc, data) {
     // Always make the bar container visible.
     doc.data.update({ displayBars: CONST.TOKEN_DISPLAY_MODES.ALWAYS });
 
-    const barConfig = game.settings.get("barbrawl", "defaultResources");
-    if (!barConfig || Object.keys(barConfig).length === 0) return;
-
     const actor = game.actors.get(data.actorId);
-    if (!actor) return;
+    if (!actor || hasProperty(actor.data, "token.flags.barbrawl.resourceBars")) return; // Don't override prototype.
 
-    if (hasProperty(actor.data, "token.flags.barbrawl.resourceBars")) return; // Do not override prototype token.
-    if (hasProperty(data, "flags.barbrawl.resourceBars")) { // Warn when overriding system defaults.
-        console.warn("barbrawl | Overriding existing resource bar configuration with user defaults.");
-    }
+    const barConfig = getDefaultResources(actor.type);
+    if (!barConfig) return;
+    doc.data.update({
+        "flags.barbrawl.resourceBars": barConfig,
+        "bar1.attribute": barConfig.bar1?.attribute ?? null,
+        "bar2.attribute": barConfig.bar2?.attribute ?? null
+    });
+});
 
-    doc.data.update({ "flags.barbrawl.resourceBars": barConfig });
+Hooks.on("preCreateActor", function (doc, data) {
+    if (!data.token || foundry.utils.hasProperty(data.token, "flags.barbrawl.resourceBars")) return;
+
+    const barConfig = getDefaultResources(doc.type);
+    if (!barConfig) return;
+    doc.data.update({
+        token: {
+            "flags.barbrawl.resourceBars": barConfig,
+            "bar1.attribute": barConfig.bar1?.attribute ?? null,
+            "bar2.attribute": barConfig.bar2?.attribute ?? null
+        }
+    });
 });
 
 /** Hook to update bar visibility. */